Fix GCC warnings

+++ prboom2/src/dsda/gl/render_scale.c
@@ -71,7 +71,7 @@ void dsda_GLSetRenderViewportParams() {
   gl_scale_y = (float)gl_viewport_height / (float)SCREENHEIGHT;
 
   // elim - This will be zero if no statusbar is being drawn
-  gl_statusbar_height = (int)ceilf(gl_scale_y * (float)ST_SCALED_HEIGHT) * R_PartialView();
+  gl_statusbar_height = ceilf(gl_scale_y * (float)ST_SCALED_HEIGHT) * R_PartialView();
 
   gl_scene_width = gl_viewport_width;
   gl_scene_height = gl_viewport_height - gl_statusbar_height;
